requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3   
certifi==2024.12.14      
charset-normalizer==3.4.0
dnspython==2.7.0
idna==3.10
logging==0.4.9.6
markdownify==0.13.1      
numpy==1.26.3
pandas==2.2.3
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jobspy==1.1.76
pytz==2024.2
regex==2024.11.6
requests==2.32.3
regex==2024.11.6
regex==2024.11.6
regex==2024.11.6
requests==2.32.3
regex==2024.11.6
requests==2.32.3
six==1.17.0
six==1.17.0
soupsieve==2.6
tls-client==1.0.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3